Move the whole contents of setup.rs in liminal.rs
Remove setup.rs. All the methods defined in setup.rs are used exclusively
in liminal.rs, except for tests, so might as well park them there, for
efficiency.
